+++ ajax/lockfields.php
@@ -1,0 +1,53 @@
+<?php
+
+/**
+ * -------------------------------------------------------------------------
+ * YAGP plugin for GLPI
+ * Copyright (C) 2019-2025 by the TICgal Team.
+ * https://tic.gal/en/project/yagp-yet-another-glpi-plugin/
+ * -------------------------------------------------------------------------
+ * LICENSE
+ * This file is part of the YAGP plugin.
+ * YAGP plugin is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 3 of the License, or
+ * (at your option) any later version.
+ * YAGP plugin is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * You should have received a copy of the GNU General Public License
+ * along with YAGP. If not, see <http://www.gnu.org/licenses/>.
+ * -------------------------------------------------------------------------
+ * @package   yagp
+ * @author    the TICGAL team
+ * @copyright Copyright (c) 2025 TICGAL team
+ * @license   AGPL License 3.0 or (at your option) any later version
+ *            http://www.gnu.org/licenses/agpl-3.0-standalone.html
+ * @link      https://www.tic.gal
+ * @since     2019
+ * -------------------------------------------------------------------------
+ */
+
+include("../../../inc/includes.php");
+header("Content-Type: application/json; charset=UTF-8");
+Html::header_nocache();
+
+$plugin = new Plugin();
+if (!$plugin->isInstalled('yagp') || !$plugin->isActivated('yagp')) {
+    Html::displayNotFoundError();
+}
+
+Session::checkLoginUser();
+
+$locked = [];
+// Never expose locked fields for anything but the logged-in user's own
+// account: an admin editing someone else's profile must not be affected.
+if ((int) ($_GET['id'] ?? 0) === (int) Session::getLoginUserID()) {
+    $config = PluginYagpConfig::getInstance();
+    if ($config->fields['lockfields']) {
+        $locked = PluginYagpProfile::getLockedFieldsForSession();
+    }
+}
+
+echo json_encode(['locked' => $locked]);