+++ inc/config.class.php
@@ -239,6 +239,8 @@ class PluginYagpConfig extends CommonDBTM
                 `default_satisfaction` INT {$default_key_sign} NOT NULL DEFAULT '3',
                 `modal_satisfaction` TINYINT(1) NOT NULL DEFAULT '0',
                 `observers_affect_status` TINYINT(1) NOT NULL DEFAULT '0',
+                `lockfields` TINYINT(1) NOT NULL DEFAULT '0',
+                `lockfields_auditlog` TINYINT(1) NOT NULL DEFAULT '0',
                 PRIMARY KEY  (`id`)
 			) ENGINE=InnoDB DEFAULT CHARSET={$default_charset}
             COLLATE={$default_collation} ROW_FORMAT=DYNAMIC;";
@@ -276,6 +278,9 @@ class PluginYagpConfig extends CommonDBTM
             $migration->addField($table, 'modal_satisfaction', 'boolean', ['value' => 0]);
             // * 2.6.0
             $migration->addField($table, 'observers_affect_status', 'boolean', ['value' => 0]);
+            // * 2.7.0
+            $migration->addField($table, 'lockfields', 'boolean', ['value' => 0]);
+            $migration->addField($table, 'lockfields_auditlog', 'boolean', ['value' => 0]);
 
             $migration->migrationOneTable($table);
         }